[CCUBE][1528]Cleaned up public types exports for font and typo
